refactor: standardize HTTP responses using ResponseEntity in Ninja and Mission controllers

+++ ninja-management-api/src/main/java/dev/backend/ninja_management_api/Mission/MissionController.java
@@ -1,5 +1,7 @@
 package dev.backend.ninja_management_api.Mission;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,27 +17,45 @@ public class MissionController {
     }
 
     @PostMapping("/create")
-    public MissionDTO create(@RequestBody MissionDTO missionDTO) {
-        return missionService.create(missionDTO);
+    public ResponseEntity<String> create(@RequestBody MissionDTO missionDTO) {
+        MissionDTO newMission = missionService.create(missionDTO);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Successfully created mission: " + newMission.getName());
     }
 
     @GetMapping("/list")
-    public List<MissionDTO> listMissions() {
-        return missionService.list();
+    public ResponseEntity<List<MissionDTO>> listMissions() {
+        List<MissionDTO> missions = missionService.list();
+        return ResponseEntity.ok(missions);
     }
 
     @GetMapping("/listById/{id}")
-    public MissionDTO listById(@PathVariable Integer id) {
-        return missionService.listById(id);
+    public ResponseEntity<MissionDTO> listById(@PathVariable Integer id) {
+        MissionDTO mission = missionService.listById(id);
+        if (mission != null) {
+            return ResponseEntity.ok(mission);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Integer id) {
-        missionService.delete(id);
+    public ResponseEntity<String> delete(@PathVariable Integer id) {
+        if (missionService.listById(id) != null) {
+            missionService.delete(id);
+            return ResponseEntity.ok("Mission with ID " + id + " successfully deleted!");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("The mission with ID " + id + " was not found!");
     }
 
     @PutMapping("/update/{id}")
-    public MissionDTO update(@PathVariable Integer id, @RequestBody MissionDTO missionDTO) {
-        return missionService.update(id, missionDTO);
+    public ResponseEntity<MissionDTO> update(@PathVariable Integer id, @RequestBody MissionDTO missionDTO) {
+        MissionDTO mission = missionService.update(id, missionDTO);
+        if (mission != null) {
+            return ResponseEntity.ok(mission);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
     }
 }
